[FIX] mail: deleting attachment with bad internet
task-2714158

closes odoo/odoo#95680

X-original-commit: 85a2b2683485a193151b02a93d20ce06b722826d

+++ addons/mail/static/src/models/attachment.js
@@ -77,20 +77,24 @@ registerModel({
             if (!this.isUploading) {
                 this.update({ isUnlinkPending: true });
                 try {
-                    await this.async(() => this.env.services.rpc({
+                    await this.env.services.rpc({
                         route: `/mail/attachment/delete`,
                         params: {
                             access_token: this.accessToken,
                             attachment_id: this.id,
                         },
-                    }, { shadow: true }));
+                    }, { shadow: true });
                 } finally {
-                    this.update({ isUnlinkPending: false });
+                    if (this.exists()) {
+                        this.update({ isUnlinkPending: false });
+                    }
                 }
             } else if (this.uploadingAbortController) {
                 this.uploadingAbortController.abort();
             }
-            this.delete();
+            if (this.exists()) {
+                this.delete();
+            }
         },
         /**
          * @private
